refactor: update API imports
